make table component
